# Refactor `GetShards` command