--- a/es.go
+++ b/es.go
@@ -873,59 +873,40 @@ func (c *Client) GetPrettyIndexMappings(index string) (string, error) {
 //
 //Use case: You can view shard information on all nodes or a subset.
 func (c *Client) GetShards(nodes []string) ([]Shard, error) {
-	var response []Shard
+	var allShards []Shard
 	req := c.buildGetRequest("_cat/shards")
-	body, err := handleErrWithBytes(req)
+	err := handleErrWithStruct(req, &allShards)
 
 	if err != nil {
 		return nil, err
 	}
 
-	d := json.NewDecoder(strings.NewReader(string(body)))
-
-	// Read opening bracket
-	_, err = d.Token()
-
-	if err != nil {
-		fmt.Printf("Error parsing JSON. Expected '['")
+	// No nodes passed, so return all shards
+	if len(nodes) == 0 {
+		return allShards, nil
 	}
 
-	// Iterate over the array elements for single pass filtering
-	for d.More() {
-		var shard Shard
-		// Get Shard object from array
-		err = d.Decode(&shard)
+	var filteredShards []Shard
+	var nodeRegexps []*regexp.Regexp
 
+	for _, node := range nodes {
+		nodeRegexp, err := regexp.Compile(node)
 		if err != nil {
 			return nil, err
 		}
+		nodeRegexps = append(nodeRegexps, nodeRegexp)
+	}
 
-		if len(nodes) == 0 {
-			// No nodes passed, so return all shards
-			response = append(response, shard)
-		} else if len(nodes) > 0 {
-			// Only return nodes of interest
-			for _, node := range nodes {
-				// Support regexp matching of node name
-				matches, err := regexp.MatchString(node, shard.Node)
+	for _, shard := range allShards {
+		for _, nodeRegexp := range nodeRegexps {
+			// Support regexp matching of node name
+			matches := nodeRegexp.MatchString(shard.Node)
 
-				if err != nil {
-					return nil, err
-				}
-
-				if matches {
-					response = append(response, shard)
-				}
+			if matches {
+				filteredShards = append(filteredShards, shard)
 			}
 		}
 	}
 
-	// Read closing bracket
-	_, err = d.Token()
-
-	if err != nil {
-		fmt.Printf("Error parsing JSON. Expected ']'")
-	}
-
-	return response, nil
+	return filteredShards, nil
 }